add PyPIConGPU documentation
Merge documentation of PyPIConGPU into docs/
Place into separate PyPIConGPU section after developer doc, rationale:
It is developer documentation, however it does not fit with the other
documents in the "Development" section, also it discusses an entirely
module.

The newly added section utilizes documentation generated by plantuml.
This generation is *NOT* part of the sphinx build process, as plantuml
is quite a heavy tool. After adjusting the developer has to manually
call `make` and commit the generated images.